refactor(test): Move backog tests to separate project
