+++ NationalArchives.FindCaseLaw.Backlog.Tests/MetadataTests/TestDelimitedArrayConverter.cs
@@ -3,9 +3,12 @@
 using System.Linq;
 
 using Backlog.Csv;
+
 using CsvHelper;
 using CsvHelper.Configuration;
+
 using Moq;
+
 using Xunit;
 
 namespace test.backlog.MetadataTests;
