+++ NationalArchives.FindCaseLaw.Backlog.Tests/MetadataTests/TestMakeMetadata.cs
@@ -24,7 +24,7 @@ public class TestMakeMetadata
         {
             id = "123",
             Court = "UKFTT-GRC",
-            DecisionDateTime = new DateTime(2023, 01, 14,  14, 30, 00, DateTimeKind.Utc),
+            DecisionDateTime = new DateTime(2023, 01, 14, 14, 30, 00, DateTimeKind.Utc),
             CaseNo = ["ABC/2023/001"],
             Claimants = "John Smith",
             Respondent = "HMRC",
@@ -71,7 +71,7 @@ public class TestMakeMetadata
             id = "124",
             FilePath = "/test/data/test.pdf",
             Court = "UKFTT-GRC",
-            DecisionDateTime = new DateTime(2023, 01, 14,  14, 30, 00, DateTimeKind.Utc),
+            DecisionDateTime = new DateTime(2023, 01, 14, 14, 30, 00, DateTimeKind.Utc),
             CaseNo = ["ABC/2023/002"],
             Appellants = "Jane Doe",
             Respondent = "Home Office",
@@ -158,7 +158,7 @@ public class TestMakeMetadata
         var line = CsvMetadataLineHelper.DummyLineWithClaimants with
         {
             Court = "UKFTT-GRC",
-            DecisionDateTime = new DateTime(2010, 01, 10,  14, 30, 00, DateTimeKind.Utc)
+            DecisionDateTime = new DateTime(2010, 01, 10, 14, 30, 00, DateTimeKind.Utc)
         };
 
         // Act
@@ -190,7 +190,7 @@ public class TestMakeMetadata
         // Arrange
         var line = CsvMetadataLineHelper.DummyLineWithClaimants with
         {
-            DecisionDateTime = new DateTime(2023, 12, 25,  15, 45, 30, DateTimeKind.Utc),
+            DecisionDateTime = new DateTime(2023, 12, 25, 15, 45, 30, DateTimeKind.Utc),
         };
 
         // Act
@@ -257,7 +257,7 @@ public class TestMakeMetadata
         // Arrange
         var line = CsvMetadataLineHelper.DummyLineWithClaimants with
         {
-           Jurisdictions = ["Transport"]
+            Jurisdictions = ["Transport"]
         };
 
         // Act
@@ -334,7 +334,8 @@ public class TestMakeMetadata
     {
         var line = CsvMetadataLineHelper.DummyLineWithClaimants with
         {
-            Ncn = null, DecisionDateTime = new DateTime(2019, 03, 21, 0, 0, 0, DateTimeKind.Utc)
+            Ncn = null,
+            DecisionDateTime = new DateTime(2019, 03, 21, 0, 0, 0, DateTimeKind.Utc)
         };
 
         var result = MetadataTransformer.MakeMetadata(line);
