+++ NationalArchives.FindCaseLaw.Backlog.Tests/MetadataTests/TestRead.cs
@@ -163,9 +163,9 @@ public sealed class TestRead : IDisposable
 
         Assert.Empty(result);
         var expectedErrorMessage = $"Line 2: Could not convert field `Uuid` with value \"{badUuid}\" to type `Guid` [{csvLine}]";
-       mockTracker.Verify(t => t.TrackCsvParseError(expectedErrorMessage), Times.Once);
+        mockTracker.Verify(t => t.TrackCsvParseError(expectedErrorMessage), Times.Once);
     }
-    
+
     [Fact]
     public void Read_WithDodgySkippedLines_DoesNotOutputValidationErrors()
     {
