+++ NationalArchives.FindCaseLaw.Backlog.Tests/NationalArchives.FindCaseLaw.Backlog.Tests.csproj
@@ -1,0 +1,30 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <AssemblyAttribute Include="Xunit.CaptureConsole" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="../backlog/backlog.csproj" />
+    <ProjectReference Include="../test/test.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="18.8.1" />
+    <PackageReference Include="Moq" Version="4.20.72" />
+    <PackageReference Include="Shouldly" Version="4.3.0" />
+    <PackageReference Include="TestableIO.System.IO.Abstractions.TestingHelpers" Version="22.2.0" />
+    <PackageReference Include="xunit.v3" Version="3.2.2" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="3.1.5" />
+
+    <EmbeddedResource Include="./expected-output/*.json" />
+    <EmbeddedResource Include="./expected-output/*.xml" />
+    <EmbeddedResource Include="./test-data/*/court_documents/*"/>
+  </ItemGroup>
+</Project>
