+++ src/Program.cs
@@ -18,6 +18,7 @@ using AkN = UK.Gov.Legislation.Judgments.AkomaNtoso;
 using Api = UK.Gov.NationalArchives.Judgments.Api;
 
 [assembly: InternalsVisibleTo("test")]
+[assembly: InternalsVisibleTo("NationalArchives.FindCaseLaw.Backlog.Tests")]
 [assembly: InternalsVisibleTo("backlog")]
 
 public class Program
